move lookup table dependece to i10_apple2
